loss utils and layer-category specific packages
