+++ deeplearning4j-modelimport/src/main/java/org/deeplearning4j/nn/modelimport/keras/layers/core/KerasFlatten.java
@@ -1,37 +1,34 @@
-package org.deeplearning4j.nn.modelimport.keras.layers;
+package org.deeplearning4j.nn.modelimport.keras.layers.core;
 
-import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.deeplearning4j.nn.conf.InputPreProcessor;
 import org.deeplearning4j.nn.conf.inputs.InputType;
-import org.deeplearning4j.nn.conf.layers.GlobalPoolingLayer;
-import org.deeplearning4j.nn.conf.preprocessor.FeedForwardToRnnPreProcessor;
+import org.deeplearning4j.nn.conf.inputs.InputType.InputTypeConvolutional;
+import org.deeplearning4j.nn.conf.preprocessor.CnnToFeedForwardPreProcessor;
+import org.deeplearning4j.nn.conf.preprocessor.RnnToFeedForwardPreProcessor;
 import org.deeplearning4j.nn.modelimport.keras.exceptions.InvalidKerasConfigurationException;
 import org.deeplearning4j.nn.modelimport.keras.KerasLayer;
 import org.deeplearning4j.nn.modelimport.keras.exceptions.UnsupportedKerasConfigurationException;
+import org.deeplearning4j.nn.modelimport.keras.preprocessors.TensorFlowCnnToFeedForwardPreProcessor;
 
 import java.util.Map;
 
 /**
- * Imports a Keras Pooling layer as a DL4J Subsampling layer.
+ * Imports a Keras Flatten layer as a DL4J {Cnn,Rnn}ToFeedForwardInputPreProcessor.
  *
  * @author dave@skymind.io
  */
 @Slf4j
-@Data
-public class KerasGlobalPooling extends KerasLayer {
-
-    private final int[] dimensions;
+public class KerasFlatten extends KerasLayer {
 
     /**
      * Constructor from parsed Keras layer configuration dictionary.
      *
-     * @param layerConfig   dictionary containing Keras layer configuration.
-     *
+     * @param layerConfig       dictionary containing Keras layer configuration
      * @throws InvalidKerasConfigurationException
      * @throws UnsupportedKerasConfigurationException
      */
-    public KerasGlobalPooling(Map<String, Object> layerConfig)
+    public KerasFlatten(Map<String, Object> layerConfig)
                     throws InvalidKerasConfigurationException, UnsupportedKerasConfigurationException {
         this(layerConfig, true);
     }
@@ -44,24 +41,19 @@ public class KerasGlobalPooling extends KerasLayer {
      * @throws InvalidKerasConfigurationException
      * @throws UnsupportedKerasConfigurationException
      */
-    public KerasGlobalPooling(Map<String, Object> layerConfig, boolean enforceTrainingConfig)
+    public KerasFlatten(Map<String, Object> layerConfig, boolean enforceTrainingConfig)
                     throws InvalidKerasConfigurationException, UnsupportedKerasConfigurationException {
         super(layerConfig, enforceTrainingConfig);
-        this.dimensions = mapPoolingDimensions(this.className);
-        GlobalPoolingLayer.Builder builder =
-                        new GlobalPoolingLayer.Builder(mapPoolingType(this.className)).poolingDimensions(dimensions)
-                                        .collapseDimensions(true).name(this.layerName).dropOut(this.dropout);
-        this.layer = builder.build();
-        this.vertex = null;
     }
 
     /**
-     * Get DL4J SubsamplingLayer.
+     * Whether this Keras layer maps to a DL4J InputPreProcessor.
      *
-     * @return  SubsamplingLayer
+     * @return      true
      */
-    public GlobalPoolingLayer getGlobalPoolingLayer() {
-        return (GlobalPoolingLayer) this.layer;
+    @Override
+    public boolean isInputPreProcessor() {
+        return true;
     }
 
     /**
@@ -72,15 +64,28 @@ public class KerasGlobalPooling extends KerasLayer {
      * @throws InvalidKerasConfigurationException
      * @see org.deeplearning4j.nn.conf.InputPreProcessor
      */
+    @Override
     public InputPreProcessor getInputPreprocessor(InputType... inputType) throws InvalidKerasConfigurationException {
         if (inputType.length > 1)
             throw new InvalidKerasConfigurationException(
-                            "Keras GlobalPooling layer accepts only one input (received " + inputType.length + ")");
-        InputPreProcessor preprocessor;
-        if (inputType[0].getType() == InputType.Type.FF && this.dimensions.length == 1) {
-            preprocessor = new FeedForwardToRnnPreProcessor();
-        } else {
-            preprocessor = this.getGlobalPoolingLayer().getPreProcessorForInputType(inputType[0]);
+                            "Keras Flatten layer accepts only one input (received " + inputType.length + ")");
+        InputPreProcessor preprocessor = null;
+        if (inputType[0] instanceof InputTypeConvolutional) {
+            InputTypeConvolutional it = (InputTypeConvolutional) inputType[0];
+            switch (this.getDimOrder()) {
+                case NONE:
+                case THEANO:
+                    preprocessor = new CnnToFeedForwardPreProcessor(it.getHeight(), it.getWidth(), it.getDepth());
+                    break;
+                case TENSORFLOW:
+                    preprocessor = new TensorFlowCnnToFeedForwardPreProcessor(it.getHeight(), it.getWidth(),
+                                    it.getDepth());
+                    break;
+                default:
+                    throw new InvalidKerasConfigurationException("Unknown Keras backend " + this.getDimOrder());
+            }
+        } else if (inputType[0] instanceof InputType.InputTypeRecurrent) {
+            preprocessor = new RnnToFeedForwardPreProcessor();
         }
         return preprocessor;
     }
@@ -96,13 +101,7 @@ public class KerasGlobalPooling extends KerasLayer {
     public InputType getOutputType(InputType... inputType) throws InvalidKerasConfigurationException {
         if (inputType.length > 1)
             throw new InvalidKerasConfigurationException(
-                            "Keras Subsampling layer accepts only one input (received " + inputType.length + ")");
-
-        /* Check whether layer requires a preprocessor for this InputType. */
-        InputPreProcessor preprocessor = getInputPreprocessor(inputType[0]);
-        if (preprocessor != null) {
-            return this.getGlobalPoolingLayer().getOutputType(-1, preprocessor.getOutputType(inputType[0]));
-        }
-        return this.getGlobalPoolingLayer().getOutputType(-1, inputType[0]);
+                            "Keras Flatten layer accepts only one input (received " + inputType.length + ")");
+        return getInputPreprocessor(inputType).getOutputType(inputType[0]);
     }
 }
