+++ deeplearning4j-modelimport/src/test/java/org/deeplearning4j/nn/modelimport/keras/KerasLayerTest.java
@@ -25,10 +25,17 @@ import org.deeplearning4j.nn.modelimport.keras.config.Keras1LayerConfiguration;
 import org.deeplearning4j.nn.modelimport.keras.config.Keras2LayerConfiguration;
 import org.deeplearning4j.nn.modelimport.keras.config.KerasLayerConfiguration;
 import org.deeplearning4j.nn.modelimport.keras.exceptions.UnsupportedKerasConfigurationException;
-import org.deeplearning4j.nn.modelimport.keras.layers.*;
 import org.deeplearning4j.nn.modelimport.keras.layers.advanced.activations.KerasLeakyReLU;
 import org.deeplearning4j.nn.modelimport.keras.layers.convolutional.KerasConvolution1D;
 import org.deeplearning4j.nn.modelimport.keras.layers.convolutional.KerasConvolution2D;
+import org.deeplearning4j.nn.modelimport.keras.layers.core.KerasActivation;
+import org.deeplearning4j.nn.modelimport.keras.layers.core.KerasDense;
+import org.deeplearning4j.nn.modelimport.keras.layers.core.KerasDropout;
+import org.deeplearning4j.nn.modelimport.keras.layers.core.KerasReshape;
+import org.deeplearning4j.nn.modelimport.keras.layers.embeddings.KerasEmbedding;
+import org.deeplearning4j.nn.modelimport.keras.layers.normalization.KerasBatchNormalization;
+import org.deeplearning4j.nn.modelimport.keras.layers.pooling.KerasPooling;
+import org.deeplearning4j.nn.modelimport.keras.layers.recurrent.KerasLstm;
 import org.deeplearning4j.nn.modelimport.keras.preprocessors.ReshapePreprocessor;
 import org.deeplearning4j.nn.weights.WeightInit;
 import org.junit.Test;
